+++ CHANGELOG.md
@@ -9,6 +9,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 
 ### Breaking Changes
 - Constructors and factory methods of `SceneReference` now validate their arguments and throw exceptions of type `SceneReferenceCreationException` if they are invalid. Note that the default constructor always creates an empty instance, but it never throws.
+- Simplified names related to GUID. All public-facing changes are as follows (old -> new):
+	- Public constructor argument name change: `SceneReference(string sceneAssetGuidHex)` -> `SceneReference(string guid)`
+	- Public property name change: `SceneReference.AssetGuidHex` -> `SceneReference.Guid`
+	- Serialized internal field name change: `SceneReference.sceneAssetGuidHex` -> `SceneReference.guid` (This change does not require you to change anything in your code. However, Unity will re-serialize your Unity-serialized `SceneReference`s as it comes across them. Please commit these re-serialization changes as you see fit, otherwise they will keep re-appearing until you do so. You will not lose any data as we utilize `FormerlySerializedAs`.)
 
 ### Added
 - `SceneReference` now supports custom XML serialization via `System.Xml`.